+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Added] `:store` option, defaults to `nil` to save memory usage
 - [Added] `:max_pages` option
+- [Added] `Crawler.running?/1` to check whether Crawler is running
 - [Improved] The queue is being supervised now
 
 ## v1.2.0 [2023-09-29]